active device pan fix, tasker filters done

+++ routes/delete.js
@@ -22,37 +22,42 @@ router.delete('/task', async (req, res) => {
 router.delete('/floor', async (req, res) => {
   const id = req.query.id;
   try {
-    const floors = await db.tx(async t => {
-      const deletedFloor = await t.one(
-        `DELETE FROM floors WHERE id = '${id}' RETURNING *`
-      );
-      const updatedFloors = await t.manyOrNone(
-        `UPDATE floors SET position = position - 1 WHERE position > ${deletedFloor.position} RETURNING *`
-      );
-      return updatedFloors;
-    });
+    const allFloors = await db.manyOrNone('SELECT * FROM floors');
+    if (allFloors.length > 1) {
+      const floors = await db.tx(async t => {
+        const deletedFloor = await t.one(
+          `DELETE FROM floors WHERE id = '${id}' RETURNING *`
+        );
+        const updatedFloors = await t.manyOrNone(
+          `UPDATE floors SET position = position - 1 WHERE position > ${deletedFloor.position} RETURNING *`
+        );
+        return updatedFloors;
+      });
 
-    const taskerQuery =
-      'SELECT t1."deviceId", t1."itemId", ' +
-      't1."isCheckedOff", devices.floor from devices ' +
-      'INNER JOIN tasker_items_devices t1 ' +
-      'ON t1."deviceId" = devices.id INNER JOIN tasker_items ' +
-      'ON t1."itemId" = tasker_items.id';
-    const [walls, areas, devices, tasks, taskDevices] = await db.multi(
-      'SELECT * FROM walls;SELECT * FROM areas;' +
-        'SELECT * FROM devices;SELECT * FROM tasker_items;' +
-        taskerQuery
-    );
+      const taskerQuery =
+        'SELECT t1."deviceId", t1."itemId", ' +
+        't1."isCheckedOff", devices.floor from devices ' +
+        'INNER JOIN tasker_items_devices t1 ' +
+        'ON t1."deviceId" = devices.id INNER JOIN tasker_items ' +
+        'ON t1."itemId" = tasker_items.id';
+      const [walls, areas, devices, tasks, taskDevices] = await db.multi(
+        'SELECT * FROM walls;SELECT * FROM areas;' +
+          'SELECT * FROM devices;SELECT * FROM tasker_items;' +
+          taskerQuery
+      );
 
-    res.json({
-      deletedFloorId: id,
-      floors,
-      walls,
-      areas,
-      devices,
-      tasks,
-      taskDevices
-    });
+      res.json({
+        deletedFloorId: id,
+        floors,
+        walls,
+        areas,
+        devices,
+        tasks,
+        taskDevices
+      });
+    } else {
+      throw new Error('Cannot delete the only floor');
+    }
   } catch (error) {
     console.log(error);
     res.status(400).json({ error: 'Server error' });
